+++ relayer/src/chain/mock.rs
@@ -145,10 +145,18 @@ impl Chain for MockChain {
 
     fn build_header(
         &self,
-        _trusted_light_block: Self::LightBlock,
-        _target_light_block: Self::LightBlock,
+        trusted_light_block: Self::LightBlock,
+        target_light_block: Self::LightBlock,
     ) -> Result<Self::Header, Error> {
-        unimplemented!()
+        Ok(Self::Header {
+            signed_header: target_light_block.signed_header.clone(),
+            validator_set: target_light_block.validators,
+            trusted_height: Height::new(
+                self.id().version(),
+                u64::from(trusted_light_block.signed_header.header.height),
+            ),
+            trusted_validator_set: trusted_light_block.validators,
+        })
     }
 
     fn query_latest_height(&self) -> Result<Height, Error> {