+++ relayer/src/foreign_client.rs
@@ -2,9 +2,6 @@ use prost_types::Any;
 use thiserror::Error;
 use tracing::info;
 
-use ibc_proto::ibc::core::client::v1::MsgCreateClient as RawMsgCreateClient;
-use ibc_proto::ibc::core::client::v1::MsgUpdateClient as RawMsgUpdateClient;
-
 use ibc::ics02_client::header::Header;
 use ibc::ics02_client::msgs::create_client::MsgCreateAnyClient;
 use ibc::ics02_client::msgs::update_client::MsgUpdateAnyClient;
@@ -13,6 +10,8 @@ use ibc::ics02_client::state::ConsensusState;
 use ibc::ics24_host::identifier::{ChainId, ClientId};
 use ibc::tx_msg::Msg;
 use ibc::Height;
+use ibc_proto::ibc::core::client::v1::MsgCreateClient as RawMsgCreateClient;
+use ibc_proto::ibc::core::client::v1::MsgUpdateClient as RawMsgUpdateClient;
 
 use crate::chain::handle::ChainHandle;
 use crate::error::{Error, Kind};
@@ -28,7 +27,10 @@ pub enum ForeignClientError {
 
 #[derive(Clone, Debug)]
 pub struct ForeignClientConfig {
+    /// The identifier of the chain which hosts this client
     chain: ChainId,
+
+    /// The client's identifier
     id: ClientId,
 }
 
@@ -49,7 +51,7 @@ impl ForeignClientConfig {
     }
 }
 
-#[derive(Clone)]
+#[derive(Clone, Debug)]
 pub struct ForeignClient {
     config: ForeignClientConfig,
 }
@@ -61,8 +63,8 @@ impl ForeignClient {
     /// TODO: what are the pre-conditions for success?
     /// Is it enough to have a "live" handle to each of `dst_chain` and `src_chain` chains?
     pub fn new(
-        dst_chain: impl ChainHandle,
-        src_chain: impl ChainHandle,
+        dst_chain: &impl ChainHandle,
+        src_chain: &impl ChainHandle,
         config: ForeignClientConfig,
     ) -> Result<ForeignClient, ForeignClientError> {
         let done = '\u{1F36D}';
@@ -131,8 +133,8 @@ impl ForeignClient {
 }
 
 pub fn build_create_client(
-    dst_chain: impl ChainHandle,
-    src_chain: impl ChainHandle,
+    dst_chain: &impl ChainHandle,
+    src_chain: &impl ChainHandle,
     dst_client_id: &ClientId,
 ) -> Result<MsgCreateAnyClient, Error> {
     // Verify that the client has not been created already, i.e the destination chain does not
@@ -164,18 +166,18 @@ pub fn build_create_client(
 }
 
 pub fn build_create_client_and_send(
-    dst_chain: impl ChainHandle,
-    src_chain: impl ChainHandle,
+    dst_chain: &impl ChainHandle,
+    src_chain: &impl ChainHandle,
     opts: &ForeignClientConfig,
 ) -> Result<Vec<String>, Error> {
-    let new_msg = build_create_client(dst_chain.clone(), src_chain, opts.client_id())?;
+    let new_msg = build_create_client(dst_chain, src_chain, opts.client_id())?;
 
     Ok(dst_chain.send_msgs(vec![new_msg.to_any::<RawMsgCreateClient>()])?)
 }
 
 pub fn build_update_client(
-    dst_chain: impl ChainHandle,
-    src_chain: impl ChainHandle,
+    dst_chain: &impl ChainHandle,
+    src_chain: &impl ChainHandle,
     dst_client_id: &ClientId,
     target_height: Height,
 ) -> Result<Vec<Any>, Error> {
@@ -199,13 +201,13 @@ pub fn build_update_client(
 }
 
 pub fn build_update_client_and_send(
-    dst_chain: impl ChainHandle,
-    src_chain: impl ChainHandle,
+    dst_chain: &impl ChainHandle,
+    src_chain: &impl ChainHandle,
     opts: &ForeignClientConfig,
 ) -> Result<Vec<String>, Error> {
     let new_msgs = build_update_client(
-        dst_chain.clone(),
-        src_chain.clone(),
+        dst_chain,
+        src_chain,
         opts.client_id(),
         src_chain.query_latest_height()?,
     )?;
@@ -213,51 +215,202 @@ pub fn build_update_client_and_send(
     Ok(dst_chain.send_msgs(new_msgs)?)
 }
 
+/// Tests the integration of crates `relayer` plus `relayer-cli` against crate `ibc`. These tests
+/// exercise various client methods (create, update, ForeignClient::new) using locally-running
+/// instances of chains built using `MockChain`.
 #[cfg(test)]
 mod test {
     use std::str::FromStr;
 
     use ibc::ics24_host::identifier::ClientId;
+    use ibc::Height;
 
+    use crate::chain::handle::ChainHandle;
     use crate::chain::mock::test_utils::get_basic_chain_config;
     use crate::chain::mock::MockChain;
     use crate::chain::runtime::ChainRuntime;
     use crate::foreign_client::{
-        build_create_client_and_send, build_update_client_and_send, ForeignClientConfig,
+        build_create_client_and_send, build_update_client_and_send, ForeignClient,
+        ForeignClientConfig,
     };
 
+    /// Basic test for the `build_create_client_and_send` method.
     #[test]
-    fn test_build_create_client_and_send() {
-        let client_id = ClientId::from_str("client_on_a_forb").unwrap();
+    fn create_client_and_send_method() {
+        let a_client_id = ClientId::from_str("client_on_a_forb").unwrap();
+        let b_client_id = ClientId::from_str("client_on_b_fora").unwrap();
         let a_cfg = get_basic_chain_config("chain_a");
         let b_cfg = get_basic_chain_config("chain_b");
-        let opts = ForeignClientConfig::new(&a_cfg.id, &client_id);
+        let a_opts = ForeignClientConfig::new(&a_cfg.id, &a_client_id);
+        let b_opts = ForeignClientConfig::new(&b_cfg.id, &b_client_id);
 
         let (a_chain, _) = ChainRuntime::<MockChain>::spawn(a_cfg).unwrap();
         let (b_chain, _) = ChainRuntime::<MockChain>::spawn(b_cfg).unwrap();
 
-        let res = build_create_client_and_send(a_chain, b_chain, &opts);
+        // Create the client on chain a
+        let res = build_create_client_and_send(&a_chain, &b_chain, &a_opts);
         assert!(
             res.is_ok(),
-            "build_create_client_and_send failed with error {:?}",
+            "build_create_client_and_send failed (chain a) with error {:?}",
+            res
+        );
+
+        // Double client creation should be forbidden.
+        let res = build_create_client_and_send(&a_chain, &b_chain, &a_opts);
+        assert!(
+            res.is_err(),
+            "build_create_client_and_send double client creation should have failed!",
+        );
+
+        // Create the client on chain b
+        let res = build_create_client_and_send(&b_chain, &a_chain, &b_opts);
+        assert!(
+            res.is_ok(),
+            "build_create_client_and_send failed (chain b) with error {:?}",
+            res
+        );
+
+        // Test double creation for chain b
+        let res = build_create_client_and_send(&b_chain, &a_chain, &b_opts);
+        assert!(
+            res.is_err(),
+            "build_create_client_and_send failed (chain b) with error {:?}",
             res
         );
     }
 
+    /// Basic test for the `build_update_client_and_send` & `build_create_client_and_send` methods.
     #[test]
-    fn test_build_update_client_and_send() {
-        let client_id = ClientId::from_str("client_on_a_forb").unwrap();
+    fn update_client_and_send_method() {
         let a_cfg = get_basic_chain_config("chain_a");
         let b_cfg = get_basic_chain_config("chain_b");
-        let opts = ForeignClientConfig::new(&a_cfg.id, &client_id);
+        let a_client_id = ClientId::from_str("client_on_a_forb").unwrap();
+        let a_opts = ForeignClientConfig::new(&a_cfg.id, &a_client_id);
+        let b_client_id = ClientId::from_str("client_on_b_fora").unwrap();
+        let b_opts = ForeignClientConfig::new(&b_cfg.id, &b_client_id);
+
+        // The number of ping-pong iterations
+        let num_iterations = 3;
 
         let (a_chain, _) = ChainRuntime::<MockChain>::spawn(a_cfg).unwrap();
         let (b_chain, _) = ChainRuntime::<MockChain>::spawn(b_cfg).unwrap();
 
-        let res = build_update_client_and_send(a_chain, b_chain, &opts);
+        // This action should fail because no client exists (yet)
+        let res = build_update_client_and_send(&a_chain, &b_chain, &a_opts);
         assert!(
             res.is_err(),
             "build_update_client_and_send was supposed to fail (no client existed)"
+        );
+
+        // Remember b's height.
+        let b_height_start = b_chain.query_latest_height().unwrap();
+
+        // Create a client on chain a
+        let res = build_create_client_and_send(&a_chain, &b_chain, &a_opts);
+        assert!(
+            res.is_ok(),
+            "build_create_client_and_send failed (chain a) with error {:?}",
+            res
+        );
+
+        // This should fail because the client on chain a already has the latest headers. Chain b,
+        // the source chain for the client on a, is at the same height where it was when the client
+        // was created, so an update should fail here.
+        let res = build_update_client_and_send(&a_chain, &b_chain, &a_opts);
+
+        assert!(
+            res.is_err(),
+            "build_update_client_and_send was supposed to fail",
+        );
+        let b_height_last = b_chain.query_latest_height().unwrap();
+        assert_eq!(b_height_last, b_height_start);
+
+        // Create a client on chain b
+        let res = build_create_client_and_send(&b_chain, &a_chain, &b_opts);
+        assert!(
+            res.is_ok(),
+            "build_create_client_and_send failed (chain b) with error {:?}",
+            res
+        );
+        // Chain b should have advanced
+        let mut b_height_last = b_chain.query_latest_height().unwrap();
+        assert_eq!(b_height_last, b_height_start.increment());
+
+        // Remember the current height of chain a
+        let mut a_height_last = a_chain.query_latest_height().unwrap();
+
+        // Now we can update both clients -- a ping pong, similar to ICS18 `client_update_ping_pong`
+        for _i in 1..num_iterations {
+            let res = build_update_client_and_send(&a_chain, &b_chain, &a_opts);
+            assert!(
+                res.is_ok(),
+                "build_update_client_and_send failed (chain a) with error: {:?}",
+                res
+            );
+            let a_height_current = a_chain.query_latest_height().unwrap();
+            a_height_last = a_height_last.increment();
+            assert_eq!(
+                a_height_last, a_height_current,
+                "after client update, chain a did not advance"
+            );
+
+            // And also update the client on chain b.
+            let res = build_update_client_and_send(&b_chain, &a_chain, &b_opts);
+            assert!(
+                res.is_ok(),
+                "build_update_client_and_send failed (chain b) with error: {:?}",
+                res
+            );
+            let b_height_current = b_chain.query_latest_height().unwrap();
+            b_height_last = b_height_last.increment();
+            assert_eq!(
+                b_height_last, b_height_current,
+                "after client update, chain b did not advance"
+            );
+        }
+    }
+
+    /// Tests for `ForeignClient::new()`.
+    #[test]
+    fn foreign_client_create() {
+        let a_cfg = get_basic_chain_config("chain_a");
+        let b_cfg = get_basic_chain_config("chain_b");
+        let a_client_id = ClientId::from_str("client_on_a_forb").unwrap();
+        let a_opts = ForeignClientConfig::new(&a_cfg.id, &a_client_id);
+        let b_client_id = ClientId::from_str("client_on_b_fora").unwrap();
+        let b_opts = ForeignClientConfig::new(&b_cfg.id, &b_client_id);
+
+        let (a_chain, _) = ChainRuntime::<MockChain>::spawn(a_cfg).unwrap();
+        let (b_chain, _) = ChainRuntime::<MockChain>::spawn(b_cfg).unwrap();
+
+        // Instantiate the foreign clients on the two chains.
+        let client_on_a = ForeignClient::new(&a_chain, &b_chain, a_opts);
+        assert!(
+            client_on_a.is_ok(),
+            "Client creation (on chain a) failed with error: {:?}",
+            client_on_a
+        );
+
+        let client_on_b = ForeignClient::new(&b_chain, &a_chain, b_opts);
+        assert!(
+            client_on_b.is_ok(),
+            "Client creation (on chain a) failed with error: {:?}",
+            client_on_b
+        );
+
+        // Now that the clients exists, we should be able to query its state
+        let b_client_state = b_chain.query_client_state(&b_client_id, Height::default());
+        assert!(
+            b_client_state.is_ok(),
+            "Client query (on chain b) failed with error: {:?}",
+            b_client_state
+        );
+
+        let a_client_state = a_chain.query_client_state(&a_client_id, Height::default());
+        assert!(
+            a_client_state.is_ok(),
+            "Client query (on chain a) failed with error: {:?}",
+            a_client_state
         );
     }
 }