+++ CHANGELOG.md
@@ -7,8 +7,11 @@
 ### IMPROVEMENTS
 
 - Mock chain (implementing IBC handlers) and integration against CLI ([#158])
+- Relayer tests for client update (ping pong) against MockChain ([#381])
+
 
 [#158]: https://github.com/informalsystems/ibc-rs/issues/158
+[#381]: https://github.com/informalsystems/ibc-rs/issues/381
 
 ## v0.0.5
 *December 2, 2020*
@@ -18,6 +21,11 @@ We now have the full-stack implementation for supporting client creation & updat
 We also consolidated our TLA+ specs into an "IBC Core TLA+ specification," and added ICS 020 spec. 
 
 Special thanks to external contributors for this release: @CharlyCst ([#347], [#419]).
+
+- [relayer-cli]
+  - Add `--all` option to `light rm` command to remove all peers for a given chain ([#431])
+
+[#431]: https://github.com/informalsystems/ibc-rs/issues/431
 
 ### FEATURES